News 엔티티 수정 1. 성능 최적화를 위한 AI 분석기능은 개인화 및 스케줄러 로직에서 제외 (크롤링만 수행) 2. 뉴스 API 접근 및 크롤링 시 병렬 처리 도입 (성능 90% 개선) 3. 페이징 객체 dataCount 타입 long으로 변경 4. 그 외 성능 최적화 및 구조 변경으로 인해 사용하지 않는 클래스 제거

+++ src/main/java/app/finup/common/dto/Page.java
@@ -15,12 +15,12 @@ import java.util.List;
 public class Page<T> {
 
     private List<T> rows;       // 페이징 결과 데이터 리스트
-    private int dataCount;      // 데이터의 총 개수 (같은 조건으로 조회된 총 데이터의 수)
+    private long dataCount;     // 데이터의 총 개수 (같은 조건으로 조회된 총 데이터의 수)
     private int pageNum;        // 현재 페이지
     private int pageSize;       // 한 페이지 당 가져올 행 개수
     private int groupSize;      // 그룹당 페이지 개수
 
-    private Page(List<T> rows, int dataCount, int pageNum, int pageSize, int groupSize) {
+    private Page(List<T> rows, long dataCount, int pageNum, int pageSize, int groupSize) {
         this.rows = rows;
         this.dataCount = dataCount;
         this.pageNum = pageNum + 1;
@@ -37,7 +37,7 @@ public class Page<T> {
      * @return 생성된 페이징 객체
      * @param <T> 페이징 데이터 타입
      */
-    public static <T> Page<T> of(List<T> rows, int dataCount, int pageNum, int pageSize) {
+    public static <T> Page<T> of(List<T> rows, long dataCount, int pageNum, int pageSize) {
         return new Page<>(rows, dataCount, pageNum, pageSize, 5);
     }
 
@@ -51,7 +51,7 @@ public class Page<T> {
      * @return 생성된 페이징 객체
      * @param <T> 페이징 데이터 타입
      */
-    public static <T> Page<T> of(List<T> rows, int dataCount, int pageNum, int pageSize, int groupSize) {
+    public static <T> Page<T> of(List<T> rows, long dataCount, int pageNum, int pageSize, int groupSize) {
         return new Page<>(rows, dataCount, pageNum, pageSize, groupSize);
     }
